backend/requirements.txt
# FastAPI
fastapi>=0.104.0
uvicorn[standard]>=0.24.0
python-multipart>=0.0.6
aiofiles>=23.2.0

# Database
sqlalchemy>=2.0.0
alembic>=1.12.0
psycopg2-binary>=2.9.0
asyncpg>=0.29.0
aiosqlite>=0.19.0

# Redis & Celery
redis>=5.0.0
celery[redis]>=5.3.0

# Authentication
python-jose[cryptography]>=3.3.0
passlib[bcrypt]>=1.7.0
bcrypt==4.0.1

# Validation
pydantic>=2.5.0
pydantic-settings>=2.1.0
email-validator>=2.1.0

# HTTP Client
httpx>=0.25.0

# Media Processing
opencv-python>=4.8.0
ffmpeg-python>=0.2.0
Pillow>=10.1.0

# ML Inference
torch>=2.1.0
torchvision>=0.16.0
torchaudio>=2.1.0
transformers>=4.35.0
timm>=0.9.0
openai-whisper>=20231117

# Audio Processing
librosa>=0.10.0
soundfile>=0.12.0

# Report Generation
openai>=1.3.0
reportlab>=4.0.0
python-docx>=1.1.0

# Utilities
python-dotenv>=1.0.0
structlog>=23.2.0

# Testing
pytest>=7.4.0
pytest-asyncio>=0.21.0
pytest-cov>=4.1.0
httpx>=0.25.0
